+++ packages/core/src/utils/index.ts
@@ -1,8 +1,4 @@
 export * from './decoratePlugins';
-export * from './getElementComponent';
-export * from './getElementDeserializer';
-export * from './getNodeDeserializer';
-export * from './getRenderElement';
 export * from './onDOMBeforeInputPlugins';
 export * from './onKeyDownPlugins';
 export * from './renderElementPlugins';
