+++ packages/core/tsconfig.build.json
@@ -1,4 +1,8 @@
 {
   "extends": "../../tsconfig.build",
+  "compilerOptions": {
+    "declarationDir": "./dist",
+    "outDir": "./dist"
+  },
   "include": ["src"]
 }
