--- packages/slate-plugins/jest.config.js
@@ -1,6 +1,0 @@
-const jestConfig = require('../../jest.config');
-
-module.exports = {
-  ...jestConfig,
-  setupFilesAfterEnv: ['<rootDir>/src/setupTests.ts'],
-};
