+++ packages/slate-plugins/qwdqd.js
@@ -1,0 +1,5 @@
+const qwdqd = require('../../jest.config');
+
+module.exports = {
+  ...qwdqd,
+};
