+++ packages/slate-plugins/src/__tests__/common/utils/getNodeDeserializer/default.spec.ts
@@ -1,4 +1,4 @@
-import { getNodeDeserializer } from '@udecode/core/src';
+import { getNodeDeserializer } from '../../../../common/utils/getNodeDeserializer';
 import { PARAGRAPH } from '../../../../elements/paragraph';
 
 const createNode = () => ({ type: PARAGRAPH });
