+++ packages/slate-plugins/src/__tests__/deserializers/deserialize-html/getDeserializer/createElement.ts
@@ -1,4 +1,4 @@
-import { getElementDeserializer } from '@udecode/core';
+import { getElementDeserializer } from '../../../../common/utils/getElementDeserializer';
 
 const createElement = () => ({ type: 'p' });
