+++ packages/slate-plugins/src/common/utils/index.ts
@@ -1,6 +1,10 @@
 export * from './createNode';
 export * from './createDocumentNode';
 export * from './escapeRegexp';
+export * from './getElementComponent';
+export * from './getElementDeserializer';
+export * from './getNodeDeserializer';
+export * from './getRenderElement';
 export * from './getPreventDefaultHandler';
 export * from './getHandler';
 export * from './isUrl';
