+++ packages/slate-plugins/src/elements/action-item/deserializeActionItem.ts
@@ -1,4 +1,5 @@
-import { DeserializeHtml, getElementDeserializer } from '@udecode/core';
+import { DeserializeHtml } from '@udecode/core';
+import { getElementDeserializer } from '../../common/utils/getElementDeserializer';
 import { ACTION_ITEM, ActionItemDeserializeOptions } from './types';
 
 export const deserializeActionItem = ({
