+++ packages/slate-plugins/src/elements/blockquote/deserializeBlockquote.ts
@@ -1,4 +1,5 @@
-import { DeserializeHtml, getElementDeserializer } from '@udecode/core';
+import { DeserializeHtml } from '@udecode/core';
+import { getElementDeserializer } from '../../common/utils/getElementDeserializer';
 import { BLOCKQUOTE, BlockquoteDeserializeOptions } from './types';
 
 export const deserializeBlockquote = ({
