+++ packages/slate-plugins/src/elements/blockquote/renderElementBlockquote.ts
@@ -1,4 +1,4 @@
-import { getRenderElement } from '@udecode/core';
+import { getRenderElement } from '../../common/utils/getRenderElement';
 import { BlockquoteElement } from './components';
 import { BLOCKQUOTE, BlockquoteRenderElementOptions } from './types';
