+++ packages/slate-plugins/src/elements/code-block/deserializeCodeBlock.ts
@@ -1,4 +1,5 @@
-import { DeserializeHtml, getElementDeserializer } from '@udecode/core';
+import { DeserializeHtml } from '@udecode/core';
+import { getElementDeserializer } from '../../common/utils/getElementDeserializer';
 import { CODE_BLOCK, CodeBlockDeserializeOptions } from './types';
 
 export const deserializeCodeBlock = ({
