+++ packages/slate-plugins/src/elements/image/deserializeImage.ts
@@ -1,4 +1,5 @@
-import { DeserializeHtml, getElementDeserializer } from '@udecode/core';
+import { DeserializeHtml } from '@udecode/core';
+import { getElementDeserializer } from '../../common/utils/getElementDeserializer';
 import { IMAGE, ImageDeserializeOptions } from './types';
 
 export const deserializeImage = ({
