+++ packages/slate-plugins/src/elements/list/deserializeList.ts
@@ -1,4 +1,5 @@
-import { DeserializeHtml, getElementDeserializer } from '@udecode/core';
+import { DeserializeHtml } from '@udecode/core';
+import { getElementDeserializer } from '../../common/utils/getElementDeserializer';
 import { ListDeserializeOptions, ListType } from './types';
 
 export const deserializeList = ({
