+++ packages/slate-plugins/src/elements/paragraph/deserializeParagraph.ts
@@ -1,4 +1,5 @@
-import { DeserializeHtml, getElementDeserializer } from '@udecode/core';
+import { DeserializeHtml } from '@udecode/core';
+import { getElementDeserializer } from '../../common/utils/getElementDeserializer';
 import { PARAGRAPH, ParagraphDeserializeOptions } from './types';
 
 export const deserializeParagraph = ({
