+++ packages/slate-plugins/src/elements/paragraph/renderElementParagraph.ts
@@ -1,4 +1,5 @@
-import { getElementComponent, getRenderElement } from '@udecode/core';
+import { getElementComponent } from '../../common/utils/getElementComponent';
+import { getRenderElement } from '../../common/utils/getRenderElement';
 import { PARAGRAPH, ParagraphRenderElementOptions } from './types';
 
 export const renderElementParagraph = ({
