+++ packages/slate-plugins/src/elements/table/renderElementTable.tsx
@@ -1,4 +1,5 @@
-import { getElementComponent, getRenderElements } from '@udecode/core';
+import { getElementComponent } from '../../common/utils/getElementComponent';
+import { getRenderElements } from '../../common/utils/getRenderElement';
 import { TableCell, TableElement } from './components';
 import { TableRenderElementOptions, TableType } from './types';
