+++ packages/slate-plugins/src/marks/bold/BoldPlugin.ts
@@ -1,12 +1,12 @@
-/**
- * Enables support for bold formatting
- */
 import { SlatePlugin } from '@udecode/core';
 import { onKeyDownMark } from '../../mark/onKeyDownMark';
 import { deserializeBold } from './deserializeBold';
 import { renderLeafBold } from './renderLeafBold';
 import { BoldPluginOptions, MARK_BOLD } from './types';
 
+/**
+ * Enables support for bold formatting
+ */
 export const BoldPlugin = (options: BoldPluginOptions = {}): SlatePlugin => ({
   renderLeaf: renderLeafBold(options),
   deserialize: deserializeBold(options),
