+++ packages/slate-plugins/tsconfig.build.json
@@ -4,5 +4,8 @@
     "declarationDir": "./dist",
     "outDir": "./dist"
   },
+  "references": [{
+    "path": "../core/tsconfig.build.json"
+  }],
   "include": ["src", "typings"]
 }
