+++ stories/elements/editable-voids/renderElementEditableVoid.ts
@@ -1,4 +1,5 @@
-import { getRenderElement, RenderElementOptions } from '@udecode/slate-plugins';
+import { RenderElementOptions } from '../../../packages/core/src/types';
+import { getRenderElement } from '../../../packages/slate-plugins/src/common/utils/getRenderElement';
 import { EditableVoidElement } from './EditableVoidElement';
 import { EDITABLE_VOID } from './types';
